// src/app/models/hist-infobas.ts
export interface HistInfobas {
    id: string;
    telefono: string;
    nombre_completo: string;
    fecha_nacimiento: Date;
    edad: number;
    genero: string;
    direccion: string;
    correo: string;
    fecha_registro: Date;

    sangre: string;
    peso: number;
    altura: number;
    alergias: string;
    enfermedades: string;

    contacto_emergencia: string;
    relacion: string;
    telefono_emergencia: string;

}